# expression: implement vectorized evaluation for `builtinCoalesceStringSig`