<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->
### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for builtinCoalesceStringSig, for #12106
### What is changed and how it works?
```
 go test -v -benchmem -bench=BenchmarkVectorizedBuiltinCompareFunc -run=BenchmarkVectorizedBuiltinCompareFunc -args 'builtinCoalesceStringSig'
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-VecBuiltinFunc-4         	 2308677	       515 ns/op	       0 B/op       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-NonVecBuiltinFunc-4      	   47698	     25309 ns/op	       0 B/op       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-VecBuiltinFunc#01-4      	   88285	     13372 ns/op	       0 B/op       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-NonVecBuiltinFunc#01-4   	   41358	     28975 ns/op	       0 B/op       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-VecBuiltinFunc#02-4      	   42958	     27560 ns/op	       0 B/op       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinCoalesceStringSig-NonVecBuiltinFunc#02-4   	   40330	     29633 ns/op	       0 B/op       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	8.991s
```
### Check List <!--REMOVE the items that are not applicable-->
Tests <!-- At least one of them must be included. -->
- Unit test