@qw4990 2019-10-21
I'm sorry that @b41sh is trying to use Golang template to generate all  vectorized `builtinCoalesceXXX` functions, so it's better to close this PR 🤦‍♂ .